refactor: ロジックとコンポーネントを分割
